добавляются отчеты по заполнению документов Добавляются стили для форм и таблиц

+++ DOcumentExamination/DocumentExamination/src/main/java/rzd/vivc/documentexamination/model/dto/documents/Document.java
@@ -12,6 +12,7 @@ import javax.persistence.Entity;
 import javax.persistence.FetchType;
 import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
+import javax.persistence.OneToMany;
 import javax.persistence.Table;
 import javax.persistence.Temporal;
 import javax.validation.constraints.NotNull;
@@ -59,12 +60,14 @@ public class Document extends BaseEntity implements Serializable {
     @ManyToOne
     @JoinColumn(name = "DOCUMENT_TYPE_ID")
     private DocumentType documentType;
-//</editor-fold>
+    
+
+//</editor-fold>epartment
 
     public Document() {
     }
 
-    public Document(int i) {
+    public Document(long i) {
         super(i);
     }
